% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{figlet}
\alias{figlet}
\title{figlet}
\usage{
figlet(arglist = arglist_get(...), input_file = NULL, output_file = NULL)
}
\arguments{
\item{arglist}{Arguments for figlet}

\item{input_file}{Text input file}

\item{output_file}{ASCII art output file}
}
\description{
Run figlet with input and output files.
}
\details{
If no input_file, will use arglist argument.
If no output_file, will print to console.
}
\examples{
library(outsider)
figlet <- module_import('figlet', repo = 'dombennett/om..figlet')
figlet('hello!')
}
